+++ database/migrations/2023_03_05_135500_create_pinjamruangans_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('pinjamruangans', function (Blueprint $table) {
             $table->id();
-            $table->bigIncrements('id_ruangan');
-            $table->bigIncrements('id_user');
+            $table->integer('id_ruangan');
+            $table->integer('id_user');
             $table->date('tgl_mulai');
             $table->time('wkt_mulai');
             $table->date('tgl_selesai');
